+++ packages/react/src/atomic/Badge/src/components/BadgeSkeleton/badgeSkeleton.types.ts
@@ -4,6 +4,11 @@ export type BadgeSkeletonProperties = Partial<
   Pick<SkeletonProps, "width" | "height" | "className">
 > & {
   /**
+   * Change the badge skeleton type between a text counter and a dot indicator.
+   * @default text
+   */
+  type?: "text" | "dot";
+  /**
    * This is an attribute used to identify a DOM node for testing purposes.
    */
   "data-testid"?: string;