+++ packages/react/src/atomic/Input/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./InputIcon";
 export * from "./InputPassword";
 export * from "./InputSearch";
 export * from "./InputSkeleton";
+export * from "./InputText";